Removing duplicate code and moving map spells in their own file
