+++ app/Components/TileExporter/MapSpells.js
@@ -1,0 +1,19 @@
+////here all maps spells are!
+//convert lat/lon to mercator style number
+function long2tile(lon,zoom) {
+  return (Math.floor((lon+180)/360*Math.pow(2,zoom)));
+}
+function lat2tile(lat,zoom)  {
+  return (Math.floor((1-Math.log(Math.tan(lat*Math.PI/180) + 1/Math.cos(lat*Math.PI/180))/Math.PI)/2 *Math.pow(2,zoom)));
+}
+
+//shold check it will work
+function tile2Lon(tileLon, zoom) {
+  return (tileLon*360/Math.pow(2,zoom)-180).toFixed(10);
+}
+
+function tile2Lat(tileLat, zoom) {
+  return ((360/Math.PI) * Math.atan(Math.pow( Math.E, (Math.PI - 2*Math.PI*tileLat/(Math.pow(2,zoom)))))-90).toFixed(10);
+}
+
+module.exports = {long2tile, lat2tile, tile2Lon, tile2Lat};
