+++ app/Components/TileExporter/PreviewMap.js
@@ -1,4 +1,5 @@
 import d3 from 'd3'
+import {tile2Lon, tile2Lat} from './MapSpells';
 
 import Key from '../../Keys'
 
@@ -176,13 +177,5 @@ var PreviewMap = (function() {
   }
 })();
 
-
-function tile2Lon(tileLon, zoom) {
-  return (tileLon*360/Math.pow(2,zoom)-180).toFixed(10);
-}
-
-function tile2Lat(tileLat, zoom) {
-  return ((360/Math.PI) * Math.atan(Math.pow( Math.E, (Math.PI - 2*Math.PI*tileLat/(Math.pow(2,zoom)))))-90).toFixed(10);
-}
 
 module.exports = PreviewMap
